third_party/blink/web_tests/external/wpt/domparsing/interfaces.any.worker-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
PASS Partial interface Element: original interface defined
PASS Partial interface Range: original interface defined
PASS DOMParser interface: existence and properties of interface object
FAIL DOMParser must be primary interface of new DOMParser() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: DOMParser is not defined"
FAIL Stringification of new DOMParser() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: DOMParser is not defined"
FAIL DOMParser interface: new DOMParser() must not have property "parseFromString" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: DOMParser is not defined"
PASS XMLSerializer interface: existence and properties of interface object
FAIL XMLSerializer must be primary interface of new XMLSerializer() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: XMLSerializer is not defined"
FAIL Stringification of new XMLSerializer() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: XMLSerializer is not defined"
FAIL XMLSerializer interface: new XMLSerializer() must not have property "serializeToString" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: XMLSerializer is not defined"
PASS Node interface: existence and properties of interface object
PASS DocumentFragment interface: existence and properties of interface object
PASS ShadowRoot interface: existence and properties of interface object
PASS Element interface: existence and properties of interface object
FAIL Element interface: document.createElement("div") must not have property "outerHTML" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: document is not defined"
FAIL Element interface: document.createElement("div") must not have property "insertAdjacentHTML" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: document is not defined"
FAIL Element interface: document.createElement("div") must not have property "innerHTML" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: document is not defined"
PASS AbstractRange interface: existence and properties of interface object
PASS Range interface: existence and properties of interface object
FAIL Range interface: new Range() must not have property "createContextualFragment" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Range is not defined"
Harness: the test ran to completion.
